+++ packages/react-core/src/components/Chip/__tests__/ChipGroup.test.tsx
@@ -7,6 +7,8 @@ import { ChipGroup } from '../index';
 import { Chip } from '../../Chip';
 import styles from '@patternfly/react-styles/css/components/Chip/chip-group';
 
+jest.mock('../../../helpers/GenerateId/GenerateId');
+
 test('chip group default', () => {
   render(
     <ChipGroup>
@@ -71,7 +73,7 @@ test('chip group has aria-labelledby attribute', () => {
       <Chip>1.1</Chip>
     </ChipGroup>
   );
-  expect(screen.getByRole('group')).toHaveAttribute('aria-labelledby', expect.stringContaining(`pf-random-id`));
+  expect(screen.getByRole('group')).toHaveAttribute('aria-labelledby', expect.stringContaining(`generated-id`));
 });
 
 test('chip group has aria-labelledby attribute set to ID value', () => {