+++ packages/react-core/src/components/SearchInput/__tests__/SearchInput.test.tsx
@@ -11,6 +11,7 @@ import badgeStyles from '@patternfly/react-styles/css/components/Badge/badge';
 import textInputGroupStyles from '@patternfly/react-styles/css/components/TextInputGroup/text-input-group';
 
 jest.mock('../../../helpers/OUIA/ouia');
+jest.mock('../../../helpers/GenerateId/GenerateId');
 
 const props = {
   onChange: jest.fn(),