+++ packages/react-charts/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 7.3.0-prerelease.6 (2024-03-26)
+
+**Note:** Version bump only for package @patternfly/react-charts
+
+# 7.3.0-prerelease.5 (2024-03-22)
+
+**Note:** Version bump only for package @patternfly/react-charts
+
 # [7.3.0-prerelease.4](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-charts@7.3.0-prerelease.3...@patternfly/react-charts@7.3.0-prerelease.4) (2024-03-05)
 
 **Note:** Version bump only for package @patternfly/react-charts