+++ packages/react-drag-drop/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 5.3.0-prerelease.22 (2024-03-29)
+
+**Note:** Version bump only for package @patternfly/react-drag-drop
+
+# [5.3.0-prerelease.21](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-drag-drop@5.3.0-prerelease.20...@patternfly/react-drag-drop@5.3.0-prerelease.21) (2024-03-28)
+
+**Note:** Version bump only for package @patternfly/react-drag-drop
+
+# [5.3.0-prerelease.20](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-drag-drop@5.3.0-prerelease.19...@patternfly/react-drag-drop@5.3.0-prerelease.20) (2024-03-28)
+
+**Note:** Version bump only for package @patternfly/react-drag-drop
+
+# [5.3.0-prerelease.19](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-drag-drop@5.3.0-prerelease.18...@patternfly/react-drag-drop@5.3.0-prerelease.19) (2024-03-27)
+
+**Note:** Version bump only for package @patternfly/react-drag-drop
+
+# [5.3.0-prerelease.18](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-drag-drop@5.3.0-prerelease.17...@patternfly/react-drag-drop@5.3.0-prerelease.18) (2024-03-27)
+
+**Note:** Version bump only for package @patternfly/react-drag-drop
+
+# [5.3.0-prerelease.17](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-drag-drop@5.3.0-prerelease.16...@patternfly/react-drag-drop@5.3.0-prerelease.17) (2024-03-27)
+
+**Note:** Version bump only for package @patternfly/react-drag-drop
+
+# [5.3.0-prerelease.16](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-drag-drop@5.3.0-prerelease.15...@patternfly/react-drag-drop@5.3.0-prerelease.16) (2024-03-26)
+
+**Note:** Version bump only for package @patternfly/react-drag-drop
+
+# 5.3.0-prerelease.15 (2024-03-26)
+
+**Note:** Version bump only for package @patternfly/react-drag-drop
+
+# 5.3.0-prerelease.14 (2024-03-26)
+
+**Note:** Version bump only for package @patternfly/react-drag-drop
+
+# [5.3.0-prerelease.13](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-drag-drop@5.3.0-prerelease.12...@patternfly/react-drag-drop@5.3.0-prerelease.13) (2024-03-25)
+
+**Note:** Version bump only for package @patternfly/react-drag-drop
+
+# [5.3.0-prerelease.12](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-drag-drop@5.3.0-prerelease.11...@patternfly/react-drag-drop@5.3.0-prerelease.12) (2024-03-22)
+
+**Note:** Version bump only for package @patternfly/react-drag-drop
+
 # 5.3.0-prerelease.11 (2024-03-20)
 
 **Note:** Version bump only for package @patternfly/react-drag-drop