+++ packages/react-icons/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 5.3.0-prerelease.3 (2024-03-26)
+
+**Note:** Version bump only for package @patternfly/react-icons
+
+# 5.3.0-prerelease.2 (2024-03-22)
+
+**Note:** Version bump only for package @patternfly/react-icons
+
 # 5.3.0-prerelease.1 (2024-02-28)
 
 ### Bug Fixes