+++ packages/react-table/CHANGELOG.md
@@ -3,6 +3,58 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.3.0-prerelease.24](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.3.0-prerelease.23...@patternfly/react-table@5.3.0-prerelease.24) (2024-03-29)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# 5.3.0-prerelease.23 (2024-03-29)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.3.0-prerelease.22](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.3.0-prerelease.21...@patternfly/react-table@5.3.0-prerelease.22) (2024-03-28)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.3.0-prerelease.21](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.3.0-prerelease.20...@patternfly/react-table@5.3.0-prerelease.21) (2024-03-28)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.3.0-prerelease.20](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.3.0-prerelease.19...@patternfly/react-table@5.3.0-prerelease.20) (2024-03-27)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.3.0-prerelease.19](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.3.0-prerelease.18...@patternfly/react-table@5.3.0-prerelease.19) (2024-03-27)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.3.0-prerelease.18](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.3.0-prerelease.17...@patternfly/react-table@5.3.0-prerelease.18) (2024-03-27)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.3.0-prerelease.17](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.3.0-prerelease.16...@patternfly/react-table@5.3.0-prerelease.17) (2024-03-26)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# 5.3.0-prerelease.16 (2024-03-26)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# 5.3.0-prerelease.15 (2024-03-26)
+
+### Bug Fixes
+
+- **Table:** added ActionsColumn prop to control close on click ([#10179](https://github.com/patternfly/patternfly-react/issues/10179)) ([cbbfcf4](https://github.com/patternfly/patternfly-react/commit/cbbfcf46caab3459c3c19af6256cafda3f5b40df))
+- **TableSortableResponsive demo:** end page number ([#10170](https://github.com/patternfly/patternfly-react/issues/10170)) ([b1b8706](https://github.com/patternfly/patternfly-react/commit/b1b8706680dfd161f06c3377830426e9b20aed94))
+- **Table:** updated a11y for empty/nontext Th components ([#10152](https://github.com/patternfly/patternfly-react/issues/10152)) ([48d8140](https://github.com/patternfly/patternfly-react/commit/48d8140d5ecd1d467b88c573cffbd86183dce898))
+
+# [5.3.0-prerelease.14](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.3.0-prerelease.13...@patternfly/react-table@5.3.0-prerelease.14) (2024-03-25)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.3.0-prerelease.13](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.3.0-prerelease.12...@patternfly/react-table@5.3.0-prerelease.13) (2024-03-22)
+
+**Note:** Version bump only for package @patternfly/react-table
+
 # [5.3.0-prerelease.12](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.3.0-prerelease.11...@patternfly/react-table@5.3.0-prerelease.12) (2024-03-20)
 
 **Note:** Version bump only for package @patternfly/react-table